+++ RELEASE_NOTES.md
@@ -39,3 +39,6 @@
 * Serailizer options now removed from Actor Host and placed in remoting
 * Updated documentation
 
+#### 0.0.10-alpha - Unreleased
+* Added actor lifecycle events - pre/post (Startup, Restart and Shutdown)
+* Fixed failing tests on mono and windows